services/backend/requirements.txt
aerich
fastapi
tortoise-orm
tortoise-orm[asyncpg]
PyJWT
redis
pytest
python-dotenv
passlib
python-multipart
email-validator
bcrypt
cryptography
uvicorn
pytest
pytest_asyncio
faker
prometheus-fastapi-instrumentator>=6.1.0